src/birdid/requirements.txt
tensorflow-cpu
boto3